+++ spinetoolbox/config.py
@@ -18,7 +18,7 @@ import sys
 LATEST_PROJECT_VERSION = 13
 
 # For the Add/Update SpineOpt wizard
-REQUIRED_SPINE_OPT_VERSION = "0.6.9"
+REQUIRED_SPINE_OPT_VERSION = "0.8.3"
 
 # Invalid characters for directory names
 # NOTE: "." is actually valid in a directory name but this is